move showSaveAsAiAsset, add advanced options tests
