+++ packages/model-serving/src/components/deploymentWizard/fields/__tests__/AvailableAiAssetsFields.test.tsx
@@ -1,13 +1,11 @@
 import React, { act } from 'react';
 import { render, screen, renderHook } from '@testing-library/react';
-import { ServingRuntimeModelType } from '@odh-dashboard/internal/types';
 import {
   availableAiAssetsFieldsSchema,
   AvailableAiAssetsFieldsComponent,
   isValidAvailableAiAssetsFieldsData,
   useAvailableAiAssetsFields,
 } from '../AvailableAiAssetsFields';
-import { mockDeploymentWizardState } from '../../../../__tests__/mockUtils';
 
 describe('AvailableAiAssetsFields', () => {
   describe('Schema validation', () => {
@@ -64,7 +62,6 @@ describe('AvailableAiAssetsFields', () => {
         <AvailableAiAssetsFieldsComponent
           data={{ saveAsAiAsset: false, useCase: '' }}
           setData={jest.fn()}
-          wizardData={mockDeploymentWizardState()}
         />,
       );
       expect(screen.getByTestId('save-as-ai-asset-checkbox')).toBeInTheDocument();
@@ -75,7 +72,6 @@ describe('AvailableAiAssetsFields', () => {
         <AvailableAiAssetsFieldsComponent
           data={{ saveAsAiAsset: true, useCase: '' }}
           setData={jest.fn()}
-          wizardData={mockDeploymentWizardState()}
         />,
       );
       expect(screen.getByTestId('save-as-ai-asset-checkbox')).toBeInTheDocument();
@@ -86,31 +82,12 @@ describe('AvailableAiAssetsFields', () => {
         <AvailableAiAssetsFieldsComponent
           data={{ saveAsAiAsset: true, useCase: 'test' }}
           setData={jest.fn()}
-          wizardData={mockDeploymentWizardState()}
         />,
       );
       expect(screen.getByTestId('save-as-ai-asset-checkbox')).toBeInTheDocument();
       expect(screen.getByTestId('save-as-ai-asset-checkbox')).toBeChecked();
       expect(screen.getByTestId('use-case-input')).toBeInTheDocument();
       expect(screen.getByTestId('use-case-input')).toHaveValue('test');
-    });
-    it('should not show the save as AiAsset checkbox if the model type is not generative', () => {
-      render(
-        <AvailableAiAssetsFieldsComponent
-          data={{ saveAsAiAsset: false, useCase: '' }}
-          setData={jest.fn()}
-          wizardData={mockDeploymentWizardState({
-            state: {
-              modelType: {
-                data: ServingRuntimeModelType.PREDICTIVE,
-                setData: jest.fn(),
-              },
-            },
-          })}
-        />,
-      );
-      expect(screen.queryByTestId('save-as-ai-asset-checkbox')).not.toBeInTheDocument();
-      expect(screen.queryByTestId('use-case-input')).not.toBeInTheDocument();
     });
   });
 });
